PZWidgetsKit // Add a namespace for refreshable desktop widgets.
