+++ Packages/PZWidgetsKit/Sources/PZWidgetsKit/BasicTypesAndConfigs/PlatformWidgetsEnum.swift
@@ -1,0 +1,6 @@
+// (c) 2024 and onwards Pizza Studio (AGPL v3.0 License or later).
+// ====================
+// This code is released under the SPDX-License-Identifier: `AGPL-3.0-or-later`.
+
+@available(watchOS, unavailable)
+public enum DesktopWidgets<RefreshIntent: WidgetRefreshIntentProtocol> {}
