+++ Packages/PZWidgetsKit/Sources/PZWidgetsKit/ViewsForDesktopWidgets/ViewComponents/MetaBlockRelated/MetaBar/MetaBarProtocol.swift
