+++ Packages/PZWidgetsKit/Sources/PZWidgetsKit/ViewsForDesktopWidgets/ViewComponents/TextViewComponents/StaminaRecoveryTimeText.swift
